mason: move the databricks CLI wrapper out of store_access
_databricks (the generic 'shell out to the databricks CLI' helper) lived in
store_access.py only because that was the first module to need it. deploy and
dev then imported it from there, coupling unrelated modules — mason dev has
nothing to do with store-access grants. Move it to a neutral databricks_cli
module that store_access, deploy, and dev all import from. Its tests move to
databricks_cli_test.py alongside it. No behavior change.
