+++ integrations/mason/src/databricks_mason/dev.py
@@ -17,6 +17,7 @@ import yaml
 
 from databricks_mason import render
 from databricks_mason.agent_project import AgentProject
+from databricks_mason.databricks_cli import _databricks
 from databricks_mason.deploy import (
     _upsert_manifest_env,
     mlflow_tracing_config,
@@ -26,7 +27,6 @@ from databricks_mason.deploy import (
 )
 from databricks_mason.errors import AgentCliError
 from databricks_mason.project_config import load_project_metadata
-from databricks_mason.store_access import _databricks
 
 # Default local port; `databricks apps run-local` listens here unless --app-port overrides it.
 _DEFAULT_APP_PORT = 8000
