+++ integrations/mason/tests/unit_tests/store_access_test.py
@@ -33,40 +33,6 @@ def test_memory_backend_targets_memory_entries():
     assert b.resource_name == "postgres-memory"  # distinct name so both stores coexist on one app
 
 
-def test_databricks_failure_hides_the_apps_subcommand(monkeypatch):
-    # Mason drives `databricks apps` as an implementation detail; a failure must be reported in
-    # Mason's terms (the `action`) and must NOT echo the raw subcommand or --profile at the customer.
-    def fake_run(cmd, **kw):
-        return types.SimpleNamespace(returncode=1, stdout="", stderr="boom from the CLI")
-
-    monkeypatch.setattr(sa.subprocess, "run", fake_run)
-    try:
-        sa._databricks(
-            ["apps", "run-local", "--prepare-environment"],
-            "e2-dogfood",
-            capture=True,
-            action="Could not start the agent locally.",
-        )
-        raise AssertionError("expected AgentCliError")
-    except sa.AgentCliError as exc:
-        assert exc.message == "Could not start the agent locally."
-        assert "apps" not in exc.message and "e2-dogfood" not in exc.message  # no leak
-        assert exc.hint == "boom from the CLI"  # underlying stderr preserved for debugging
-
-
-def test_databricks_failure_without_action_is_generic(monkeypatch):
-    # No action label: still no raw subcommand/profile echo — just a generic message.
-    def fake_run(cmd, **kw):
-        return types.SimpleNamespace(returncode=1, stdout="", stderr="err")
-
-    monkeypatch.setattr(sa.subprocess, "run", fake_run)
-    try:
-        sa._databricks(["apps", "deploy", "x"], "prof", capture=True)
-        raise AssertionError("expected AgentCliError")
-    except sa.AgentCliError as exc:
-        assert "apps" not in exc.message and "prof" not in exc.message
-
-
 def test_apply_postgres_resources_sends_all_backends_in_one_update(monkeypatch):
     captured = {}
